Calculate goal progress
Goal progress is a float between 0.0 and 1.0
that represents the ratio of total number of key results
with completed status divided by the total number of key results.

Examples:

If a goal has 2 key results and both are “completed” the progress is 1
If a goal has 2 key results and 1 is “completed” the progress is 0.5
If a goal has 2 key results and none of them are completed the progress
is 0

The progress is also considered zero percent, if a goal does not have any key
result associated with it.